pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "endstone-primebds"
version = "3.0.0"
description = "An essentials plugin for diagnostics, stability, and quality of life on Minecraft Bedrock Edition."
author = "PrimeStrat"
dependencies = [
    "bedrock-protocol-packets==0.0.6",
    "RapidNBT==1.1.5",
    "chest_form_api_endstone >= 1.0.0"
]

[project.entry-points."endstone"]
primebds = "endstone_primebds:PrimeBDS"

